Rename the `engines` module to `engine`
