+++ lib/src/api/engine/any/mod.rs
@@ -7,7 +7,7 @@
 //! use serde_json::json;
 //! use std::borrow::Cow;
 //! use surrealdb::sql;
-//! use surrealdb::engines::any::connect;
+//! use surrealdb::engine::any::connect;
 //! use surrealdb::opt::auth::Root;
 //!
 //! #[derive(Serialize, Deserialize)]
@@ -232,7 +232,7 @@ impl Surreal<Any> {
 	///
 	/// ```no_run
 	/// use surrealdb::Surreal;
-	/// use surrealdb::engines::any::Any;
+	/// use surrealdb::engine::any::Any;
 	///
 	/// static DB: Surreal<Any> = Surreal::init();
 	///
@@ -258,7 +258,7 @@ impl Surreal<Any> {
 /// # Examples
 ///
 /// ```no_run
-/// use surrealdb::engines::any::connect;
+/// use surrealdb::engine::any::connect;
 ///
 /// # #[tokio::main]
 /// # async fn main() -> surrealdb::Result<()> {
