+++ lib/src/api/engine/remote/ws/native.rs
@@ -6,10 +6,10 @@ use crate::api::conn::Method;
 use crate::api::conn::Param;
 use crate::api::conn::Route;
 use crate::api::conn::Router;
-use crate::api::engines::remote::ws::Client;
-use crate::api::engines::remote::ws::Response;
-use crate::api::engines::remote::ws::PING_INTERVAL;
-use crate::api::engines::remote::ws::PING_METHOD;
+use crate::api::engine::remote::ws::Client;
+use crate::api::engine::remote::ws::Response;
+use crate::api::engine::remote::ws::PING_INTERVAL;
+use crate::api::engine::remote::ws::PING_METHOD;
 use crate::api::err::Error;
 use crate::api::opt::from_value;
 use crate::api::opt::Endpoint;
