+++ lib/src/api/opt/endpoint/fdb.rs
@@ -1,5 +1,5 @@
-use crate::api::engines::local::Db;
-use crate::api::engines::local::FDb;
+use crate::api::engine::local::Db;
+use crate::api::engine::local::FDb;
 use crate::api::err::Error;
 use crate::api::opt::Endpoint;
 use crate::api::opt::IntoEndpoint;
