+++ lib/src/api/opt/endpoint/indxdb.rs
@@ -1,5 +1,5 @@
-use crate::api::engines::local::Db;
-use crate::api::engines::local::IndxDb;
+use crate::api::engine::local::Db;
+use crate::api::engine::local::IndxDb;
 use crate::api::err::Error;
 use crate::api::opt::Endpoint;
 use crate::api::opt::IntoEndpoint;
