+++ lib/src/api/opt/endpoint/mem.rs
@@ -1,5 +1,5 @@
-use crate::api::engines::local::Db;
-use crate::api::engines::local::Mem;
+use crate::api::engine::local::Db;
+use crate::api::engine::local::Mem;
 use crate::api::opt::Endpoint;
 use crate::api::opt::IntoEndpoint;
 use crate::api::opt::Strict;
