+++ lib/src/api/opt/endpoint/ws.rs
@@ -1,6 +1,6 @@
-use crate::api::engines::remote::ws::Client;
-use crate::api::engines::remote::ws::Ws;
-use crate::api::engines::remote::ws::Wss;
+use crate::api::engine::remote::ws::Client;
+use crate::api::engine::remote::ws::Ws;
+use crate::api::engine::remote::ws::Wss;
 use crate::api::err::Error;
 use crate::api::opt::Endpoint;
 use crate::api::opt::IntoEndpoint;
